net: move netaddress.
