+++ lib/bcoin-browser.js
@@ -75,6 +75,7 @@ bcoin.net = require('./net');
 bcoin.bip150 = require('./net/bip150');
 bcoin.bip151 = require('./net/bip151');
 bcoin.bip152 = require('./net/bip152');
+bcoin.netaddress = require('./net/netaddress');
 bcoin.packets = require('./net/packets');
 bcoin.peer = require('./net/peer');
 bcoin.pool = require('./net/pool');
@@ -95,7 +96,6 @@ bcoin.invitem = require('./primitives/invitem');
 bcoin.keyring = require('./primitives/keyring');
 bcoin.merkleblock = require('./primitives/merkleblock');
 bcoin.mtx = require('./primitives/mtx');
-bcoin.netaddress = require('./primitives/netaddress');
 bcoin.outpoint = require('./primitives/outpoint');
 bcoin.output = require('./primitives/output');
 bcoin.tx = require('./primitives/tx');
