+++ lib/bcoin.js
@@ -110,6 +110,7 @@ bcoin.define('net', './net');
 bcoin.define('bip150', './net/bip150');
 bcoin.define('bip151', './net/bip151');
 bcoin.define('bip152', './net/bip152');
+bcoin.define('netaddress', './net/netaddress');
 bcoin.define('packets', './net/packets');
 bcoin.define('peer', './net/peer');
 bcoin.define('pool', './net/pool');
@@ -130,7 +131,6 @@ bcoin.define('invitem', './primitives/invitem');
 bcoin.define('keyring', './primitives/keyring');
 bcoin.define('merkleblock', './primitives/merkleblock');
 bcoin.define('mtx', './primitives/mtx');
-bcoin.define('netaddress', './primitives/netaddress');
 bcoin.define('outpoint', './primitives/outpoint');
 bcoin.define('output', './primitives/output');
 bcoin.define('tx', './primitives/tx');
