+++ lib/net/hostlist.js
@@ -15,10 +15,15 @@ const Logger = require('blgr');
 const {murmur3} = require('bfilter');
 const util = require('../utils/util');
 const Network = require('../protocol/network');
-const NetAddress = require('../primitives/netaddress');
+const NetAddress = require('./netaddress');
 const List = require('../utils/list');
 const common = require('./common');
 const seeds = require('./seeds');
+
+/*
+ * Constants
+ */
+
 const POOL32 = Buffer.allocUnsafe(32);
 
 /**
@@ -67,13 +72,11 @@ class HostList {
   init() {
     const options = this.options;
     const scores = HostList.scores;
-    const hosts = IP.getPublic();
-    const port = this.address.port;
 
-    for (let i = 0; i < this.options.maxBuckets; i++)
+    for (let i = 0; i < options.maxBuckets; i++)
       this.fresh.push(new Map());
 
-    for (let i = 0; i < this.options.maxBuckets; i++)
+    for (let i = 0; i < options.maxBuckets; i++)
       this.used.push(new List());
 
     this.setSeeds(options.seeds);
@@ -81,6 +84,9 @@ class HostList {
 
     this.pushLocal(this.address, scores.MANUAL);
     this.addLocal(options.host, options.port, scores.BIND);
+
+    const hosts = IP.getPublic();
+    const port = this.address.port;
 
     for (const host of hosts)
       this.addLocal(host, port, scores.IF);
@@ -355,18 +361,19 @@ class HostList {
       return null;
 
     const now = this.network.now();
+
     let factor = 1;
 
     for (;;) {
-      let index = random(buckets.length);
-      const bucket = buckets[index];
+      const i = random(buckets.length);
+      const bucket = buckets[i];
 
       if (bucket.size === 0)
         continue;
 
-      index = random(bucket.size);
-
+      let index = random(bucket.size);
       let entry;
+
       if (buckets === this.used) {
         entry = bucket.head;
         while (index--)
@@ -375,7 +382,7 @@ class HostList {
         for (entry of bucket.values()) {
           if (index === 0)
             break;
-          index--;
+          index -= 1;
         }
       }
 
@@ -431,7 +438,6 @@ class HostList {
     let entry = this.map.get(addr.hostname);
 
     if (entry) {
-      const now = this.network.now();
       let penalty = 2 * 60 * 60;
       let interval = 24 * 60 * 60;
 
@@ -445,6 +451,7 @@ class HostList {
       entry.addr.services >>>= 0;
 
       // Online?
+      const now = this.network.now();
       if (now - addr.time < 24 * 60 * 60)
         interval = 60 * 60;
 
@@ -490,7 +497,7 @@ class HostList {
 
       entry = new HostEntry(addr, src);
 
-      this.totalFresh++;
+      this.totalFresh += 1;
     }
 
     const bucket = this.freshBucket(entry);
@@ -502,7 +509,7 @@ class HostList {
       this.evictFresh(bucket);
 
     bucket.set(entry.key(), entry);
-    entry.refCount++;
+    entry.refCount += 1;
 
     this.map.set(entry.key(), entry);
     this.needsFlush = true;
@@ -524,7 +531,7 @@ class HostList {
 
         if (--entry.refCount === 0) {
           this.map.delete(entry.key());
-          this.totalFresh--;
+          this.totalFresh -= 1;
         }
 
         continue;
@@ -546,7 +553,7 @@ class HostList {
 
     if (--old.refCount === 0) {
       this.map.delete(old.key());
-      this.totalFresh--;
+      this.totalFresh -= 1;
     }
   }
 
@@ -605,7 +612,7 @@ class HostList {
       for (const bucket of this.used) {
         if (bucket.head === head) {
           bucket.remove(entry);
-          this.totalUsed--;
+          this.totalUsed -= 1;
           head = null;
           break;
         }
@@ -615,10 +622,10 @@ class HostList {
     } else {
       for (const bucket of this.fresh) {
         if (bucket.delete(entry.key()))
-          entry.refCount--;
+          entry.refCount -= 1;
       }
 
-      this.totalFresh--;
+      this.totalFresh -= 1;
       assert(entry.refCount === 0);
     }
 
@@ -639,7 +646,7 @@ class HostList {
     if (!entry)
       return;
 
-    entry.attempts++;
+    entry.attempts += 1;
     entry.lastAttempt = now;
   }
 
@@ -686,17 +693,17 @@ class HostList {
     assert(entry.refCount > 0);
 
     // Remove from fresh.
-    let old;
+    let old = null;
     for (const bucket of this.fresh) {
       if (bucket.delete(entry.key())) {
-        entry.refCount--;
+        entry.refCount -= 1;
         old = bucket;
       }
     }
 
     assert(old);
     assert(entry.refCount === 0);
-    this.totalFresh--;
+    this.totalFresh -= 1;
 
     // Find room in used bucket.
     const bucket = this.usedBucket(entry);
@@ -704,12 +711,13 @@ class HostList {
     if (bucket.size < this.options.maxEntries) {
       entry.used = true;
       bucket.push(entry);
-      this.totalUsed++;
+      this.totalUsed += 1;
       return;
     }
 
     // No room. Evict.
     const evicted = this.evictUsed(bucket);
+
     let fresh = this.freshBucket(evicted);
 
     // Move to entry's old bucket if no room.
@@ -724,8 +732,8 @@ class HostList {
     evicted.used = false;
     fresh.set(evicted.key(), evicted);
     assert(evicted.refCount === 0);
-    evicted.refCount++;
-    this.totalFresh++;
+    evicted.refCount += 1;
+    this.totalFresh += 1;
   }
 
   /**
@@ -937,7 +945,7 @@ class HostList {
     if (!local)
       return false;
 
-    local.score++;
+    local.score += 1;
 
     return true;
   }
@@ -954,7 +962,7 @@ class HostList {
     for (const seed of this.dnsSeeds)
       jobs.push(this.populateSeed(seed));
 
-    await Promise.all(jobs);
+    return Promise.all(jobs);
   }
 
   /**
@@ -969,7 +977,7 @@ class HostList {
     for (const node of this.dnsNodes)
       jobs.push(this.populateNode(node));
 
-    await Promise.all(jobs);
+    return Promise.all(jobs);
   }
 
   /**
@@ -1078,10 +1086,11 @@ class HostList {
   fromJSON(json) {
     const sources = new Map();
     const map = new Map();
-    let totalFresh = 0;
-    let totalUsed = 0;
     const fresh = [];
     const used = [];
+
+    let totalFresh = 0;
+    let totalUsed = 0;
 
     assert(json && typeof json === 'object');
 
@@ -1092,6 +1101,7 @@ class HostList {
 
     for (const addr of json.addrs) {
       const entry = HostEntry.fromJSON(addr, this.network);
+
       let src = sources.get(entry.src.hostname);
 
       // Save some memory.
@@ -1116,8 +1126,8 @@ class HostList {
         const entry = map.get(key);
         assert(entry);
         if (entry.refCount === 0)
-          totalFresh++;
-        entry.refCount++;
+          totalFresh += 1;
+        entry.refCount += 1;
         bucket.set(key, entry);
       }
 
@@ -1143,7 +1153,7 @@ class HostList {
         assert(entry.refCount === 0);
         assert(!entry.used);
         entry.used = true;
-        totalUsed++;
+        totalUsed += 1;
         bucket.push(entry);
       }
