+++ lib/net/index.js
@@ -17,6 +17,7 @@ exports.common = require('./common');
 exports.external = require('./external');
 exports.Framer = require('./framer');
 exports.HostList = require('./hostlist');
+exports.NetAddress = require('./netaddress');
 exports.packets = require('./packets');
 exports.Parser = require('./parser');
 exports.Peer = require('./peer');
