+++ lib/net/netaddress.js
@@ -9,14 +9,14 @@
 const assert = require('assert');
 const bio = require('bufio');
 const IP = require('binet');
-const common = require('../net/common');
 const Network = require('../protocol/network');
 const util = require('../utils/util');
+const common = require('./common');
 
 /**
  * Net Address
  * Represents a network address.
- * @alias module:primitives.NetAddress
+ * @alias module:net.NetAddress
  * @property {Host} host
  * @property {Number} port
  * @property {Number} services
