+++ lib/net/packets.js
@@ -17,7 +17,7 @@ const {BloomFilter} = require('bfilter');
 const common = require('./common');
 const util = require('../utils/util');
 const bip152 = require('./bip152');
-const NetAddress = require('../primitives/netaddress');
+const NetAddress = require('./netaddress');
 const Headers = require('../primitives/headers');
 const InvItem = require('../primitives/invitem');
 const MemBlock = require('../primitives/memblock');
