+++ lib/net/peer.js
@@ -26,7 +26,7 @@ const BIP150 = require('./bip150');
 const BIP152 = require('./bip152');
 const Block = require('../primitives/block');
 const TX = require('../primitives/tx');
-const NetAddress = require('../primitives/netaddress');
+const NetAddress = require('./netaddress');
 const Network = require('../protocol/network');
 const services = common.services;
 const invTypes = InvItem.types;
