+++ lib/net/pool.js
@@ -3329,6 +3329,7 @@ class Pool extends EventEmitter {
       throw new Error('Peer is destroyed (getdata).');
 
     let now = Date.now();
+
     const items = [];
 
     for (const hash of hashes) {
