+++ lib/node/rpc.js
@@ -19,7 +19,7 @@ const secp256k1 = require('bcrypto/lib/secp256k1');
 const util = require('../utils/util');
 const common = require('../blockchain/common');
 const Amount = require('../btc/amount');
-const NetAddress = require('../primitives/netaddress');
+const NetAddress = require('../net/netaddress');
 const Script = require('../script/script');
 const Address = require('../primitives/address');
 const Block = require('../primitives/block');
