+++ lib/primitives/index.js
@@ -21,7 +21,6 @@ exports.KeyRing = require('./keyring');
 exports.MemBlock = require('./memblock');
 exports.MerkleBlock = require('./merkleblock');
 exports.MTX = require('./mtx');
-exports.NetAddress = require('./netaddress');
 exports.Outpoint = require('./outpoint');
 exports.Output = require('./output');
 exports.TX = require('./tx');
