+++ test/protocol-test.js
@@ -7,7 +7,7 @@
 const assert = require('./util/assert');
 const Network = require('../lib/protocol/network');
 const util = require('../lib/utils/util');
-const NetAddress = require('../lib/primitives/netaddress');
+const NetAddress = require('../lib/net/netaddress');
 const TX = require('../lib/primitives/tx');
 const Framer = require('../lib/net/framer');
 const Parser = require('../lib/net/parser');
